system helper: Pass the callers pid
We add the pid as OBJECT_PID to our journal entries,
which causes journald to add number of useful fields,
such as the user and binary on whose behalf we are
acting.

The api for doing this is currently a bit awkward:
we attach the pid as object data to the FlatpakDir
object. This may warrant a better solution.