system-helper: Add GetChildRepoForPull D-Bus method
There occurs a situation in --system installation where free disk space
should be atleast 2x of the size of the pull(for it to succeed). Apart
from the double disk space requirement of the pull, there is a lot of
I/O activity involved while pulling a system child repo into the parent
one. Therefore, this D-Bus method tries to address these issues by
creating a child repo at the place (inside repo/tmp) from where pulling
into the parent is just a matter of hardlinking (instead of copy). This
also makes the pull disk-space efficient.

This method primarily tries to create a $REPO/tmp/flatpak-cache-$REF-XXXXXX
directory containing a OSTree child repo. It is bind-mounted with "-bindfs"
suffixed directory using bindfs which additionally enforces relevant file-ownership
and file-creation policies. The bindfs-mounted path of the child repo is returned
back to the client for pulling the repo from the remote. The client sees the
bind mounted directory as writable (because bindfs enforced relevant permissions)
and starts the pull. As the pull proceeds, actual files that are being created on
the disk have the ownership of the mounter(i.e. in this case, root) but
as long as the bind mount exists, the client(non-privileged) is able to write to it.

After the pull is completed, we unmount the bind mount just before Deploy so the
client would not be able to modify or write to it. This now enables the helper
to verify and hardlink the child repo(as it's owned by root) into the system one[1].

If bindfs is not found on the system, we fallback on the existing codepath of
doing a --system pull (that involves copying each object).

In case of pull interruption, this method can find the partially pulled
child repo in repo/tmp, verify the child repo and re-use the cache
directory to pull the remaining objects.

Each ongoing pull is tracked using a helper struct FlatpakSystemHelperOngoingPull.
A GPtrArray ongoing_pulls keeps the track of all ongoing pulls in
the system helper.

[1] https://github.com/ostreedev/ostree/pull/1776

https://github.com/ostreedev/ostree/issues/1723