system-helper: Add CancelPull D-Bus method
If there is a pull failure in a child repo created by GetChildRepoForPull
system helper method, there is no way to go back to the system helper and
notify it to cleanup. Therefore, CancelPull is required on the pull failure
error path, so that the ongoing pull can be cleaned up nicely and prevent
any dangling mounts and subprocesses. We probably want to reuse the partial
pull in future so we don't delete the pull cache directory.

Some future work is needed to specify the expiration or explicit cleanup
of these stale pulls directories. At one end we want to be disk space
efficient and at another end, we do not want to waste user's network
bandwidth.

https://github.com/ostreedev/ostree/issues/1723