table printer: Add cell and row apis
Add a way to change individual cells and  a getter for the
current row. Also add a variation of the print() method that
returns the number of terminal rows that was generated
when formatting to given width, while skipping a number
of rows at the beginning. This method will be used in the
following commit to print a part of the table into a terminal
window of a given size.

Closes: #2371
Approved by: alexlarsson