portal: Use the new --cwd option
The first argument of the Spawn portal call is the cwd
for the command to use inside the new sandbox. The portal
was mistakenly using it as cwd when running flatpak
on the host, which leads to errors. Fix this by
passing the cwd to flatpak via the new --cwdd option.

Closes: #2418