cli: Limit length of progress bar
This limits the witdh of the progress bar, because it looked really
weird on a wide terminal and kept changing width as the text at the
end (speed, time left, etc) changed.

It is now never more than 20 chars wide, which is fine with the
new increase in fidelity using unicode blocks.