info: Fix unicode formatting issues
printf(), and by extension, g_print() interprets
field width in terms of bytes, not characters.
Therefore, using %*s for padding strings with
non-ASCII characters does not work. Instead, use
the new utility function for aligned lines.

While we are doing this, also fix up the alignment of
the extension sections in the output.

Closes: #2502
Approved by: alexlarsson