completion: Don't let --help interrupt us
We never want g_option_context_parse to produce help
output when we are in the middle of completion.
But we still want to see --help, -h, -? among the
completions.

To achieve that, disable the built-in help and add
(ignored) help entries to our option context when
we are completing.

Closes: #2585
Approved by: alexlarsson