dir: Don't allow collection ID collisions
With Flatpak you should only have to trust each remote to provide good
updates for the apps provided by it. However the P2P support in flatpak
considers each remote to be equally trustworthy, which opens a possible
attack vector. For example if I have a flathub remote configured and
apps installed from it and I also have a remote "sketchy-remote"
configured which I have one app installed from, I expect the Flathub
apps to update from Flathub (or to update from LAN/USB sources with
Flathub GPG signatures) and not from the sketchy-remote.

The way this attack would play out is that the sketchy-remote would
deploy the same collection ID as the victim remote (in this case
org.flathub.Stable) in order to serve updates for it. So this commit
disallows any operation on remotes that would leave two with the same
collection ID but differing GPG keys. The case of two remotes having
matching collection IDs and GPG keys is allowed for e.g. mirrors. This
is implemented in flatpak_dir_modify_remote() so that it applies to both
adding new remotes and modifying existing ones, which can happen via
the CLI, via library API, or by setting metadata keys on the server
side.

Unfortunately this implementation is not perfect, because there's not
always a one-to-one mapping between configured remotes and GPG keyrings.
On Endless OS some remotes have keyrings in /usr/share/keyrings/ rather
than /var/lib/flatpak/repo/remote_name.trustedkeys.gpg as do remotes
added by Flatpak. Therefore it might make sense to have other
mitigations for this issue beyond this commit.

Partial fix for https://github.com/flatpak/flatpak/issues/1447